Move stash entry to models
